# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  description: |
    API для маркетплейса с OpenAPI-first разработкой.
    Все DTO и серверные интерфейсы генерируются из этого контракта.
servers:
  - url: /
...
paths:
...
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Обновление access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access and refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/RefreshTokenInvalidError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                error_code: VALIDATION_ERROR
                message: Validation failed
                details:
                  fields:
                    - field: price
                      violation: must be greater than 0
...
    RefreshTokenInvalidError:
      description: Invalid refresh token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: REFRESH_TOKEN_INVALID
            message: Refresh token invalid
            details: null
...
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - PRODUCT_NOT_FOUND
        - PRODUCT_INACTIVE
        - ORDER_NOT_FOUND
        - ORDER_LIMIT_EXCEEDED
        - ORDER_HAS_ACTIVE
        - INVALID_STATE_TRANSITION
        - INSUFFICIENT_STOCK
        - PROMO_CODE_INVALID
        - PROMO_CODE_MIN_AMOUNT
        - ORDER_OWNERSHIP_VIOLATION
        - VALIDATION_ERROR
        - ACCESS_DENIED
        - TOKEN_EXPIRED
        - TOKEN_INVALID
        - REFRESH_TOKEN_INVALID
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          minLength: 16
          maxLength: 2048
...
    AuthTokensResponse:
      type: object
      required: [access_token, refresh_token, token_type, access_expires_at, refresh_expires_at]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        access_expires_at:
          type: string
          format: date-time
        refresh_expires_at:
          type: string
          format: date-time
...